Finished the new script. Commit before the big split.

+++ torchtrainer/data/reference_dataset.py
@@ -136,7 +136,7 @@ class BaseSegmentationTransform:
         img_out, mask_out = self.transforms(img_tv, mask_tv)
 
         img_out = img_out.float()
-        mask_out = mask_out.float()
+        mask_out = mask_out.long()
 
         return img_out, mask_out
